+++ Chargeable/Framework/Integrations/ItemExtensionsIntegration.cs
@@ -1,0 +1,13 @@
+﻿namespace DaLion.Chargeable.Framework.Integrations;
+
+#region using directives
+
+using DaLion.Shared.Attributes;
+using DaLion.Shared.Integrations;
+
+#endregion using directives
+
+/// <summary>Initializes a new instance of the <see cref="ItemExtensionsIntegration"/> class.</summary>
+[ModRequirement("mistyspring.ItemExtensions")]
+internal sealed class ItemExtensionsIntegration()
+    : ModIntegration<ItemExtensionsIntegration, IItemExtensionsApi>(ModHelper.ModRegistry);