+++ Professions/Framework/Patchers/Prestige/Game1ActiveClickbleMenuSetterPatcher.cs
@@ -24,17 +24,22 @@ internal sealed class Game1ActiveClickbleMenuSetterPatcher : HarmonyPatcher
 
     /// <summary>Reload profession sprites on level-up.</summary>
     [HarmonyPostfix]
-    private static void FarmerStaminaSetterPostfix(IClickableMenu value)
+    private static void Game1ActiveClickbleMenuSetterPostfix(IClickableMenu? value)
     {
-        if (value is not LevelUpMenu { isProfessionChooser: true } levelup)
+        switch (value)
         {
-            return;
-        }
+            case LevelUpMenu { isProfessionChooser: true } levelup:
+                var level = Reflector.GetUnboundFieldGetter<LevelUpMenu, int>("currentLevel").Invoke(levelup);
+                if (level > 10)
+                {
+                    ModHelper.GameContent.InvalidateCacheAndLocalized("LooseSprites/Cursors");
+                }
 
-        var level = Reflector.GetUnboundFieldGetter<LevelUpMenu, int>("currentLevel").Invoke(levelup);
-        if (level > 10)
-        {
-            ModHelper.GameContent.InvalidateCacheAndLocalized("LooseSprites/Cursors");
+                break;
+            case GameMenu:
+            case null:
+                ModHelper.GameContent.InvalidateCacheAndLocalized("LooseSprites/Cursors");
+                break;
         }
     }