+++ Professions/Framework/Patchers/Prestige/LevelUpMenuUpdatePatcher.cs
@@ -83,7 +83,7 @@ internal sealed class LevelUpMenuUpdatePatcher : HarmonyPatcher
                         break;
                     case 20:
                         var rootId = player.GetCurrentRootProfessionForSkill(skill);
-                        IProfession root = Profession.FromValue(rootId);
+                        IProfession root = Profession.FromValue(rootId - 100);
                         ___professionsToChoose.AddRange(root.GetBranchingProfessions
                             .Select(p => p.Id)
                             .Where(player.professions.Contains));