+++ Shared/Extensions/Stardew/BuildingExtensions.cs
@@ -91,7 +91,7 @@ public static class BuildingExtensions
     /// </summary>
     /// <typeparam name="T">The target type.</typeparam>
     /// <param name="building">The <see cref="Building"/>.</param>
-    /// <param name="candidates">The candidate <see cref="Building"/>s, if already available.</param>
+    /// <param name="candidates">The candidate <see cref="T"/>s, if already available.</param>
     /// <param name="getPosition">A delegate to retrieve the tile coordinates of <typeparamref name="T"/>.</param>
     /// <param name="distance">The actual tile distance to the closest candidate found.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
@@ -130,7 +130,7 @@ public static class BuildingExtensions
     /// </summary>
     /// <typeparam name="TBuilding">A subtype of <see cref="Building"/>.</typeparam>
     /// <param name="building">The <see cref="Building"/>.</param>
-    /// <param name="candidates">The candidate <see cref="Building"/>s, if already available.</param>
+    /// <param name="candidates">The candidate <see cref="TBuilding"/>s, if already available.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
     /// <returns>The <see cref="Building"/> of type <typeparamref name="TBuilding"/> with the minimal distance to <paramref name="building"/>.</returns>
     /// <remarks>
@@ -158,7 +158,7 @@ public static class BuildingExtensions
     /// </summary>
     /// <typeparam name="TCharacter">A subtype of <see cref="Character"/>.</typeparam>
     /// <param name="building">The <see cref="Building"/>.</param>
-    /// <param name="candidates">The candidate <see cref="Character"/>s, if already available.</param>
+    /// <param name="candidates">The candidate <see cref="TCharacter"/>s, if already available.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
     /// <returns>The <see cref="Character"/> of type <typeparamref name="TCharacter"/> with the minimal distance to <paramref name="building"/>.</returns>
     /// <remarks>
@@ -202,7 +202,7 @@ public static class BuildingExtensions
     /// </summary>
     /// <typeparam name="TObject">A subtype of <see cref="SObject"/>.</typeparam>
     /// <param name="building">The <see cref="Building"/>.</param>
-    /// <param name="candidates">The candidate <see cref="SObject"/>s, if already available.</param>
+    /// <param name="candidates">The candidate <see cref="TObject"/>s, if already available.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
     /// <returns>The <see cref="SObject"/> of type <typeparamref name="TObject"/> with the minimal distance to <paramref name="building"/>.</returns>
     /// <remarks>
@@ -225,7 +225,7 @@ public static class BuildingExtensions
     /// </summary>
     /// <typeparam name="TTerrainFeature">A subtype of <see cref="TerrainFeature"/>.</typeparam>
     /// <param name="building">The <see cref="Building"/>.</param>
-    /// <param name="candidates">The candidate <see cref="TerrainFeature"/>, if already available.</param>
+    /// <param name="candidates">The candidate <see cref="TTerrainFeature"/>, if already available.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
     /// <returns>The <see cref="TerrainFeature"/> of type <typeparamref name="TTerrainFeature"/> with the minimal distance to <paramref name="building"/>.</returns>
     /// <remarks>