+++ Shared/Extensions/Stardew/CharacterExtensions.cs
@@ -72,7 +72,7 @@ public static class CharacterExtensions
     /// </summary>
     /// <typeparam name="T">The target type.</typeparam>
     /// <param name="character">The <see cref="Character"/>.</param>
-    /// <param name="candidates">The candidate <see cref="Building"/>s, if already available.</param>
+    /// <param name="candidates">The candidate <see cref="T"/>s, if already available.</param>
     /// <param name="getPosition">A delegate to retrieve the pixel coordinates of <typeparamref name="T"/>.</param>
     /// <param name="distance">The actual tile distance to the closest candidate found.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
@@ -111,7 +111,7 @@ public static class CharacterExtensions
     /// </summary>
     /// <typeparam name="TBuilding">A subtype of <see cref="Building"/>.</typeparam>
     /// <param name="character">The <see cref="Character"/>.</param>
-    /// <param name="candidates">The candidate <see cref="Building"/>s, if already available.</param>
+    /// <param name="candidates">The candidate <see cref="TBuilding"/>s, if already available.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
     /// <returns>The <see cref="Building"/> of type <typeparamref name="TBuilding"/> with the minimal distance to <paramref name="character"/>.</returns>
     public static TBuilding? GetClosestBuilding<TBuilding>(
@@ -130,7 +130,7 @@ public static class CharacterExtensions
     /// </summary>
     /// <typeparam name="TCharacter">A subtype of <see cref="Character"/>.</typeparam>
     /// <param name="character">The <see cref="Character"/>.</param>
-    /// <param name="candidates">The candidate <see cref="Character"/>s, if already available.</param>
+    /// <param name="candidates">The candidate <see cref="TCharacter"/>s, if already available.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
     /// <returns>The <see cref="Character"/> of type <typeparamref name="TCharacter"/> with the minimal distance to <paramref name="character"/>.</returns>
     public static TCharacter? GetClosestCharacter<TCharacter>(
@@ -177,7 +177,7 @@ public static class CharacterExtensions
     /// </summary>
     /// <typeparam name="TObject">A subtype of <see cref="SObject"/>.</typeparam>
     /// <param name="character">The <see cref="Character"/>.</param>
-    /// <param name="candidates">The candidate <see cref="SObject"/>s, if already available.</param>
+    /// <param name="candidates">The candidate <see cref="TObject"/>s, if already available.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
     /// <returns>The <see cref="SObject"/> of type <typeparamref name="TObject"/> with the minimal distance to <paramref name="character"/>.</returns>
     public static TObject? GetClosestObject<TObject>(
@@ -196,7 +196,7 @@ public static class CharacterExtensions
     /// </summary>
     /// <typeparam name="TTerrainFeature">A subtype of <see cref="TerrainFeature"/>.</typeparam>
     /// <param name="character">The <see cref="Character"/>.</param>
-    /// <param name="candidates">The candidate <see cref="TerrainFeature"/>s if already available.</param>
+    /// <param name="candidates">The candidate <see cref="TTerrainFeature"/>s if already available.</param>
     /// <param name="predicate">An optional condition with which to filter out candidates.</param>
     /// <returns>The <see cref="TerrainFeature"/> of type <typeparamref name="TTerrainFeature"/> with the minimal distance to <paramref name="character"/>.</returns>
     public static TTerrainFeature? GetClosestTerrainFeature<TTerrainFeature>(