+++ Core/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Removed Ring recipe patchers. Apparently that was added in 1.6. Also removed the Glow/Magnet Ring recipe changes.
 
+<sup><sup>[🔼 Back to top](#core-changelog)</sup></sup>
 
 ## 1.0.0 - Initial 1.6 release